+++ Clients/config/ns-inet.sh
@@ -66,8 +66,8 @@ ip netns add $NS
 echo "Waiting for APs (10 secs)"
 sleep 10 # wait for AP docker
 
-# Add WiFi interfaces wlan 40-59
-for I in `seq 40 59` ; do
+# Add WiFi interfaces wlan 40-69
+for I in `seq 40 69` ; do
 	PHY=`ls /sys/class/ieee80211/*/device/net/ | grep -B1 wlan$I | grep -Eo 'phy[0-9]+'`
 	iw phy $PHY set netns name /run/netns/$NS
 done